Add action, components, implement yt api route
Add action function to submit new contest
Extract code from createContest to new components
Rename contest.js to entry.js
Implement yt api route in backend to fetch video title & check if id is correct
