+++ backend/routes/index.js
@@ -1,4 +1,5 @@
 module.exports = {
   authRoute: require('./authRoute'),
   contestRoute: require('./contestRoute'),
+  youtubeApiRoute: require('./youtubeApiRoute'),
 };
