+++ frontend/src/components/contest/ContestItem.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/prop-types */
-import { getCoverImageUrl } from "../../utils/contest";
+import { getCoverImageUrl } from "../../utils/entry";
 import EmbedVideo from "./EmbedVideo";
 import { twMerge as tm } from "tailwind-merge";
